Code-release corrections for reproducibility
This commit fixes some discrepancies between the released code
and the reported results---caused due to a couple of things that were missed out
in the effort to clean up and publish the code.

 1. The (fixed) scales for the individual-modality likelihoods were transferred
    incorrectly---should have been `softplus(0.1)` instead of 0.1
 2. We didn't include the trained Fasttext embeddings used to compute CCA and
    cross/joint coherence scores.

To expand a bit on (2), for measuring cross/joint coherence on `cub` , we use
off-the-shelf ResNets for the `cub-image` data, but train Fasttext embeddings for
`cub-sentence` data---since it's vocabulary is quite different from what is
typically used to train Fasttext.

We then use these embeddings (ResNet, trained Fasttext) to compute CCA on the
ground-truth image and sentence training data, and use the learnt embeddings to
compute the correlation for generated samples from our model.

The learnt embeddings however, can vary quite a bit due to the limited dataset
size. The embeddings we used to report results in the paper was not saved with
our models, so re-computing them as part of the analyses can result in different
numeric values including for the baseline.

Note: The relative performance of our model against the baseline remains the
same, just that the numbers can be be different.

To mitigate this, we have done a quick search for the Fasttext embeddings that
produce the same results on the baseline as reported in the paper, and
re-computed the CCA and cross/joint coherence scores on our model with this.

With these two fixes, the results from the code match those in the paper (with
even improved cross-coherence scores on `cub`).

Fixes: iffsid/mmvae:#2